more windows casts added

+++ sql/backends/monet5/bin_partition.c
@@ -59,9 +59,9 @@ pp_nr_slices(sql_rel *rel)
 	if (est < PP_MIN_SIZE)
 		nr_slices = 1;
 	else if (est/GDKnr_threads < PP_MIN_SIZE)
-		nr_slices = est/PP_MIN_SIZE;
+		nr_slices = (int)(est/PP_MIN_SIZE);
 	else
-	    nr_slices =	est/PP_MAX_SIZE;
+	    nr_slices =	(int)(est/PP_MAX_SIZE);
 	FORCEMITODEBUG
 	if (nr_slices < GDKnr_threads)
 		nr_slices = GDKnr_threads;
